Added file inclusion trick in case eval() is not available
Some servers don't have eval enabled and perhaps does not offer the user the option to turn it on. Instead of making it impossible for these users to use SuperClosure this does a file inclusion trick that basically works the same way that eval does, but without relying on eval().